+++ app/src/main/java/com/stacktivity/yandeximagesearchengine/base/BaseImageViewHolder.kt
@@ -15,14 +15,4 @@ internal abstract class BaseImageViewHolder(itemView: View): ViewHolder(itemView
         return calculateViewWidth(viewWidth, imageHeight, imageWidth)
     }
 
-    @ExperimentalUnsignedTypes
-    protected fun getGifSize(buffer: ByteBuffer): Pair<Int, Int> {
-        val bArray = ByteArray(12)
-        buffer.get(bArray, 0, 11)
-        buffer.rewind()
-        val width = bArray[6].toUByte() + (bArray[7].toUInt() shl 8)
-        val height = bArray[8].toUByte() + (bArray[9].toUInt() shl 8)
-
-        return Pair(width.toInt(), height.toInt())
-    }
 }
